api/requirements.txt
flask==3.0.1
markdown==3.5.2
python-frontmatter==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dateutil==2.8.2
